moved InvalidTokenException to common-exception module
