+++ softgroup/basis/common-exceptions/src/main/java/com/softgroup/common/exceptions/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package com.softgroup.common.jwt.exceptions;
+package com.softgroup.common.exceptions;
 
 /**
  * @author AlexKazmirchuk
