+++ softgroup/basis/common-jwt/pom.xml
@@ -17,6 +17,10 @@
             <artifactId>jjwt</artifactId>
             <version>0.7.0</version>
         </dependency>
+        <dependency>
+            <groupId>com.softgroup</groupId>
+            <artifactId>common-exceptions</artifactId>
+        </dependency>
     </dependencies>
 
 </project>
